+++ pkg/streamingpromql/operators/aggregations/sum.go
@@ -21,7 +21,7 @@ type SumAggregationGroup struct {
 	floatCompensatingValues []float64 // Compensation value for Kahan summation.
 	floatPresent            []bool
 	histogramSums           []*histogram.FloatHistogram
-	histogramPointCount     int
+	histogramPointCount     int64
 }
 
 func (g *SumAggregationGroup) AccumulateSeries(data types.InstantVectorSeriesData, timeRange types.QueryTimeRange, memoryConsumptionTracker *limiting.MemoryConsumptionTracker, emitAnnotationFunc types.EmitAnnotationFunc) error {
@@ -127,7 +127,7 @@ func (g *SumAggregationGroup) accumulateHistograms(data types.InstantVectorSerie
 // It also takes the opportunity whilst looping through the floats to check if there
 // is a conflicting Histogram present. If both are present, an empty vector should
 // be returned. So this method removes the float+histogram where they conflict.
-func (g *SumAggregationGroup) reconcileAndCountFloatPoints() (int, bool) {
+func (g *SumAggregationGroup) reconcileAndCountFloatPoints() (int64, bool) {
 	// It would be possible to calculate the number of points when constructing
 	// the series groups. However, it requires checking each point at each input
 	// series which is more costly than looping again here and just checking each
@@ -136,7 +136,7 @@ func (g *SumAggregationGroup) reconcileAndCountFloatPoints() (int, bool) {
 	// We also take two different approaches here: One with extra checks if we
 	// have both Floats and Histograms present, and one without these checks
 	// so we don't have to do it at every point.
-	floatPointCount := 0
+	floatPointCount := int64(0)
 	haveMixedFloatsAndHistograms := false
 	if len(g.floatPresent) > 0 && len(g.histogramSums) > 0 {
 		for idx, present := range g.floatPresent {
